WIP - setting up prereqs
